Third Way to Constants
FLAT_VISITOR: This is a third way to get semantically meaning access to constants in a constants class. The first is to inherit. The second is to use "static access" (e.g. {CLASS}.my_feature). The third way is to do as I have done in this class: Provide a once'd instance of the constants class with a semantically meaningful name (i.e. like `const') and then calls like `const.first_pass' now make better sense when reading. One does not have to reason about inheritance or deal with the clutter of static access references, which can be very messy!